Improve version mappings a lot, other improvements
AssetsManager: throw FileNotFoundException if asset does not exist instead of NullPointerException
EntityInformation: Remove max health
Rename some constants into caps with underline and not camel case
Version: Fix loading pre flattening mapping (mappings are now the latest not flattened version and not the first flattened version)
Remove CustomMapping.java. Use VersionMapping with parent ability (nesting)
Almost rewrite VersionMapping. Code is way better and easier to understand
Allow VersionMapping per version and inherit block and registries mapping from PRE_FLATTENING_VERSION (if needed): Custom Entity mappings per version now possible

+++ src/main/java/de/bixilon/minosoft/protocol/protocol/OutByteBuffer.java
@@ -193,7 +193,7 @@ public class OutByteBuffer {
                 writeShort((short) -1);
                 return;
             }
-            writeShort((short) connection.getMapping().getItemId(slot.getItem()));
+            writeShort((short) (int) connection.getMapping().getItemId(slot.getItem()));
             writeByte((byte) slot.getItemCount());
             writeShort(slot.getItemMetadata());
             writeNBT(slot.getNbt(connection.getMapping()));
